refactor: move all error protocols to a single file
